--- dashboard/src/components/Routes/index.tsx
@@ -1,3 +1,0 @@
-import Routes from "./Routes";
-
-export default Routes;
